🎉 v1.1.0 | Redbox API + Perks
Welcome to the first GitHub commit for this project!

This project, designed for the Redbox, is a sophisticated API powered by Express.js to regain functionality for your kiosk. This project mainly focuses on the Redbox Perks loyalty system, allowing users to create accounts and earn points for renting/purchasing DVDs.